modified the heat charm to accomodate the workers-mutiplier option
Heat charm was missing this functionality and this patch is to add
this feature. We were trying to deploy this charm in a lxc
container on a host machine with 72 cores. In end up creating 144
process with heat-engine and heat-api.

This patch will provide the option to limit the number of worker
process in case of high number of cores in a host.

Enable it for icehouse, kilo, liberty and mitaka.

Change-Id: I811f0a91e1ff9a6d95e80a4a40be063d7d047c7e
(cherry picked from commit 93045f31241997d6fddc466d854626b030e72336)